+++ R/wrcc_loadYear.R
@@ -8,6 +8,7 @@
 #' @param meta Tibble of WRCC station metadata.
 #' @param year Year to access station data for.
 #' @param forceDownload Logical flag stating whether or not to download and override existing data.
+#' @param password Password required for access to archival data.
 #' @param baseUrl Base URL for data queries.
 #' @param verbose Logical flag controlling detailed progress statements.
 #'
@@ -23,7 +24,12 @@
 #' setRawsDataDir("~/Data/RAWS/")
 #'
 #' stationMeta <- wrcc_loadMeta(stateCode = "WA")
-#' rawsObject <- wrcc_load(wrccID = "waWENU", meta = stationMeta, year = 2020)
+#' rawsObject <- wrcc_loadYear(
+#'   wrccID = "waWENU", 
+#'   meta = stationMeta, 
+#'   year = 2020,
+#'   password = "MY_PASSWORD"
+#' )
 #' 
 #' dplyr::glimpse(rawsObject)
 #' }
@@ -32,11 +38,12 @@
 #' @seealso \code{setRawsDataDir}
 #' @references \href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
 
-wrcc_load <- function(
+wrcc_loadYear <- function(
   wrccID = NULL,
   meta = NULL,
   year = NULL,
   forceDownload = FALSE,
+  password = NULL,
   baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl",
   verbose = TRUE
 ) {
@@ -74,8 +81,26 @@ wrcc_load <- function(
       message(paste("Downloading and saving data to", filePath, "."))
     }
     
+    # NOTE:  Extend start and end UTC dates by one day to capture full days in
+    # NOTE:  every timezone.
+    
+    startdate <- 
+      lubridate::now("UTC") %>% 
+      lubridate::floor_date(unit = "year") - lubridate::ddays(1)
+    
+    enddate <- 
+      lubridate::now("UTC") %>% 
+      lubridate::ceiling_date(unit = "year") + lubridate::ddays(1)
+    
     # If local data does not exist, download and return it.
-    rawsObject <- wrcc_createTimeseriesObject(wrccID = wrccID, meta = meta, baseUrl = baseUrl)
+    rawsObject <- wrcc_createTimeseriesObject(
+      wrccID = wrccID, 
+      meta = meta, 
+      startdate = strftime(startdate, "%Y%m%d%H", tz = "UTC"),
+      enddate = strftime(enddate, "%Y%m%d%H", tz = "UTC"),
+      password = password,
+      baseUrl = baseUrl
+    )
     
     # Temp solution until raws_filter~() functions are created
     rawsObject$data <- rawsObject$data %>% dplyr::filter(stringr::str_sub(.data$datetime, 0, 4) == year)
@@ -88,5 +113,25 @@ wrcc_load <- function(
   # ----- Return ---------------------------------------------------------------
   
   return(rawsObject)
+  
+}
+
+# ===== DEBUGGING ==============================================================
+
+if ( FALSE ) {
+  
+  library(RAWSmet)
+
+  setRawsDataDir("~/Data/RAWS/")
+
+  meta <- wrcc_loadMeta(stateCode = "WA")
+  
+  wrccID = "waWENU"
+  ###meta = NULL
+  year = 2020
+  forceDownload = TRUE
+  password = NULL
+  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
+  verbose = TRUE
   
 }